src/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.6
django-allauth==0.44.0
django-bootstrap-modal-forms==2.1.0
django-chartjs==2.2.1
django-countries==7.0
django-dotenv==1.4.2
django-environ==0.4.5
django-smart-selects==1.5.9
django-utils-six==2.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==2.10
install==1.3.4
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.16.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
